Problem 2.4 Reversing a vector. Language: C++
Write a program which fills a vector with the integer values from 1 to 30. Then add the value 5 at
the end of the vector. Reverse the vector using the reverse() function from the algorithm library
and print the content of the vector on the standard output using an iterator. Then replace all
occurrences of the value 5 by the value 129 using the replace() function from the algorithm
library and print the modified vector again on the standard output.